requirements.txt
Django
django-appconf==1.0.2
django-bootstrap4
django-cors-headers
django-crispy-forms
django-extensions==2.1.5
django-heroku==0.3.1
djangorestframework
djangorestframework-simplejwt==4.1.0
Faker==1.0.2
gunicorn==19.7.0
ipython==6.5.0
psycopg2-binary
pytz
Pillow
requests
whitenoise